checks 'DIN' font availability
The "Alte DIN 1451 Mittelschrift" font may not be installed in the
system. If not found, `AddText` function returns a None object,
which could not be manupilated in later steps. The script fall backs
to "Arial" by the check font function.

Systems that has the DIN font will have no effect.